refactor: extract markdown wrapping helper
- move shared wrap logic into src/ui/markdown_wrap.rs and re-export from markdown.rs
- add focused wrap unit test and note follow-up span metadata refactor in WISHLIST
